Describe security attestation (closes #72) (#89)
(some related quotes from the CRA)
> (21) In order to support and facilitate the due diligence of
manufacturers that integrate free and open-source software components
that are not subject to the essential cybersecurity requirements set out
in this Regulation into their products with digital elements, the
Commission should be able to establish voluntary security attestation
programmes, either by a delegated act supplementing this Regulation or
by requesting a European cybersecurity certification scheme pursuant to
Article 48 of Regulation (EU) 2019/881 that takes into account the
specificities of the free and open-source software development models.
The security attestation programmes should be conceived in such a way
that not only natural or legal persons developing or contributing to the
development of a product with digital elements qualifying as free and
open-source software can initiate or finance a security attestation but
also third parties, such as manufacturers that integrate such products
into their own products with digital elements, users, or Union and
national public administrations.

> Furthermore, power to adopt acts should be delegated to the Commission
to establish voluntary security attestation programmes for assessing the
conformity of products with digital elements qualifying as free and
open-source software with all or certain essential cybersecurity
requirements or other obligations laid down in this Regulation, as well
as to specify the minimum content of theEU declaration of conformity and
to supplement the elements to be included in the technical
documentation.

> In order to facilitate the due diligence obligation set out in Article
13(5), in particular as regards manufacturers that integrate free and
open-source software components in their products with digital elements,
the Commission is empowered to adopt delegated acts in accordance with
Article 61 to supplement this Regulation by establishing voluntary
security attestation programmes allowing the developers or users of
products with digital elements qualifying as free and open-source
software as well as other third parties to assess the conformity of such
products with all or certain essential cybersecurity requirements or
other obligations laid down in this Regulation.