Clarify difference between harmonised and referenced standard (#188)
- I wasn't sure if or how I should add all the references
- I wasn't sure if I should add a second question instead: One
explaining what a harmonised standard is, another one explaining what a
referenced standard is + presumption of conformity

This is switching to the relatively common format of one sentence per
line which makes diffs easier in the future.
My IDE also complains that the "name" attribute is deprecated and should
be replaced by "id". Happy to include that here if you like.

---------

Signed-off-by: Tobie Langel <tobie@unlockopen.com>
Signed-off-by: Lars Francke <lars.francke@stackable.tech>
Signed-off-by: Luis Villa <luis.villa@sonarsource.com>
Co-authored-by: Tobie Langel <tobie@unlockopen.com>