fix CRA text links (#258)
There were a variety of internal links to the CRA text that either were
broken or were incomplete (assumed a templating that wasn't there). I
tried to clean them all up.

Should be low risk but committing to a branch and doing a PR instead of
directly committing because it's been a while.

Signed-off-by: Luis Villa <luis@lu.is>